kis-daily-report/requirements.txt
# ===== CrewAI 핵심 =====
crewai==1.14.2
crewai-tools

# ===== LLM / LangChain =====
langchain
langchain-openai
langchain-community
openai

# ===== 검색 도구 =====
duckduckgo-search

# ===== 환경 변수 =====
python-dotenv

# ===== HTTP 요청 =====
requests
